% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc_remotes.R
\name{bioc_remotes}
\alias{bioc_remotes}
\alias{add_bioc_remote}
\alias{set_bioc_remotes}
\title{Convenience functions to set, rename, or validate a repo's remotes}
\usage{
add_bioc_remote(repo = ".")

set_bioc_remotes(repo = ".", org = "Bioconductor")
}
\arguments{
\item{repo}{\code{character(1)} The local path to the git repository whose
upstream remote should be set}

\item{org}{\code{character(1)} The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}
}
\value{
\itemize{
\item \code{add_bioc_remote}: adds an 'upstream' remote with the Bioconductor git
address for a given package.
\item \code{set_bioc_remotes}: updates the remotes for a given package, setting the
'origin' remote to the GitHub location and the 'upstream' remote to the
Bioconductor git server.
}
}
\description{
The function \code{add_bioc_remote} adds an 'upstream' remote that
points to \code{git@git.bioconductor.org}. If an \code{upstream} remote already
exists, it will be validated against the SSH URL. By design, the
Bioconductor remote location is called the 'upstream' remote. The 'origin'
is set to the GitHub location.
}
\details{
The \code{set_bioc_remotes} function will update a repository's remotes
by setting \code{origin} to the GitHub location and \code{upstream} to the Bioconductor
git server. If the \code{origin} remote is not set to the GitHub location, the
function will overwrite the URL. If the \code{upstream} remote is not set, the
function will add it. If the \code{upstream} remote is set, the function will
overwrite the URL with the Bioconductor SSH URL.
Note that the remotes will follow this template:

\if{html}{\out{<div class="sourceCode">}}\preformatted{origin git@github.com:\{org\}/\{package\}
upstream git@git.bioconductor.org:packages/\{package\}
}\if{html}{\out{</div>}}
}
