% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_devel_branch.R
\name{branch_all_packages}
\alias{branch_all_packages}
\title{Convenience function to create the devel branch for all GitHub packages}
\usage{
branch_all_packages(
  version = BiocManager::version(),
  old_branches = c(.OLD_DEFAULT_BRANCH, "main"),
  org = "Bioconductor",
  set_upstream = c("origin/devel", "upstream/devel"),
  clone = TRUE
)
}
\arguments{
\item{version}{character(1) The current development version of Bioconductor
given by \code{BiocManager::version()} (default). It is used to obtain the
packages currently available in the Bioconductor CRAN-like repository with
\code{BiocManager::repositories()}.}

\item{old_branches}{character() A vector of default branch names to be
replaced, by default this includes 'master' and 'main'}

\item{org}{character(1L) The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{set_upstream}{character(1) The remote location that will be tracked by
the local branch, either "origin/devel" (default) or "upstream/devel"}

\item{clone}{logical(1) Whether to clone the GitHub repository into the
current working directory (default: TRUE)}
}
\description{
This function identifies an organization's repositories that are packages
given the current version of Bioconductor (from \code{BiocManager::version()})
and identifies which repositories need to have a \code{devel} branch added.
It then adds the \code{devel} branch using the \code{create_devel_branch} function.
It is highly recommended that the user run this on the devel version of
Bioconductor to avoid missing packages that are only in devel.
}
\details{
Note that the \code{clone} argument allows the user to clone the repository first
from GitHub via SSH. It is recommended that this be enabled and that the
user running this function can clone packages via SSH and have access to
modifying packages on the GitHub organization.
}
