% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{clone_and_push_git_repo}
\alias{clone_and_push_git_repo}
\title{Clone and update a GitHub repository.}
\usage{
clone_and_push_git_repo(
  package_name,
  release = "RELEASE_3_16",
  gh_branch = "master",
  bioc_branch = "master"
)
}
\arguments{
\item{package_name}{character(1) The name of the package to be cloned and
updated}

\item{release}{character(1) The Bioconductor version tag, e.g.,
"RELEASE_3_16"}
}
\description{
This function assumes that you have admin push access to the
bioconductor github organization.
}
