% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_devel_branch.R
\name{create_devel_branch}
\alias{create_devel_branch}
\title{Create the 'devel' branch locally and on GitHub}
\usage{
create_devel_branch(
  package_name,
  from_branch = .OLD_DEFAULT_BRANCH,
  org = "Bioconductor",
  set_upstream = c("origin/devel", "upstream/devel"),
  clone = FALSE,
  is_package = TRUE
)
}
\arguments{
\item{package_name}{character(1) The name of the organization R package that
is also available on GitHub.}

\item{from_branch}{character(1) The old default branch from which to base the
new 'devel' branch from (default: 'master')}

\item{org}{character(1L) The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{set_upstream}{character(1) The remote location that will be tracked by
the local branch, either "origin/devel" (default) or "upstream/devel"}

\item{clone}{logical(1) Whether to clone the GitHub repository into the
current working directory (default: TRUE)}

\item{is_package}{logical(1) Whether the repository is an R package on
Bioconductor. If so, additional validity checks will be run on the git
remotes.}
}
\value{
Called for the side effect of creating a 'devel' branch on the local
and remote repositories on GitHub
}
\description{
The function is meant to be run one level up from the local git repository.
It will create the 'devel' branch and push to the \code{origin} remote which
should be set to GitHub. Upstream tracking can be configured to either the
\code{origin} or \code{upstream} remote.
}
\details{
The \code{origin} remote is assumed to be GitHub, i.e.,
\code{git@github.com:user/package} but this requirement is not checked or
enforced; thus, allowing flexibility in remote \code{origin} locations. The
\code{upstream} remote is validated against the Bioconductor git repository
address, i.e., \code{git@git.bioconductor.org:packages/package}. The local
repository is validated before the \code{devel} branch is created.
}
\examples{
if (interactive()) {

  create_devel_branch(
    package_name = "SummarizedExperiment",
    org = "Bioconductor",
    set_upstream = "upstream/devel"
  )

}

}
