% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{.github_userpwd}
\alias{.github_userpwd}
\title{Get API keys from location}
\usage{
.github_userpwd()
}
\description{
Most important step of all of the code! Without getting access to
the API keys, it is impossible to run this code.
}
