% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackagesNEWS.R
\name{extractNEWS}
\alias{extractNEWS}
\title{Extract NEWS files from source package tarballs}
\usage{
extractNEWS(
  reposRoot,
  srcContrib,
  destDir = file.path(reposRoot, "news"),
  unpackDir = tempdir()
)
}
\arguments{
\item{reposRoot}{Top level path for CRAN-style repos}

\item{srcContrib}{Location of source packages}

\item{destDir}{where to extract}

\item{unpackDir}{character(1) The directory to extract files to (default is
the \code{tempdir()} location)}
}
\description{
Extracts NEWS files from source tarballs of packages.
}
\keyword{utilities}
