% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_release_gh.R
\name{get-bioc-packages}
\alias{get-bioc-packages}
\alias{get_bioc_software_manifest}
\alias{get_org_packages}
\title{Get the Bioconductor packages in the software manifest}
\usage{
get_bioc_software_manifest()

get_org_packages(version, org, type)
}
\arguments{
\item{version}{\code{character(1)} The numeric version (as character) of the
Bioconductor release, e.g., "3.16".}

\item{org}{\code{character(1)} The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{type}{\code{character()} The repository names to look through as returned by
\code{BiocManager::repositories()}. Currently, only software and experiment data
('BioCsoft' and 'BioCexp', respectively) are supported.}
}
\value{
\itemize{
\item \code{get_bioc_software_manifest}: A character vector of package names
\item \code{get_org_packages}: A named character vector of package names and their
default branches
}
}
\description{
To ensure that a GitHub repository is a software package, its
name is checked against a list of Bioconductor packages. This list is
called the \code{manifest}. The \code{get_bioc_software_manifest} function obtains
the manifest using \code{git}. Alternatively, \code{get_org_packages} uses
\code{available.packages()} with Bioconductor repositories (\code{repos=...}).
}
\examples{
get_bioc_software_manifest()

get_org_packages(
   version = bioc_version_yaml(),
   org = "Bioconductor",
   type = c("BioCsoft", "BioCexp")
)
}
