% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{get_bioc_github_repos}
\alias{get_bioc_github_repos}
\title{Get list of packages for Bioconductor's github organization.}
\usage{
get_bioc_github_repos()
}
\description{
This uses the github API, to get the bioconductor packages hosted
on github.  Return all packages in https:://github.com/Bioconductor
}
