% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{get_org_github_repos}
\alias{get_org_github_repos}
\title{Get all repositories for a given GitHub organization}
\usage{
get_org_github_repos(per_page = 100, pages = 10, org = "Bioconductor")
}
\arguments{
\item{per_page}{Number of items to return per page. If omitted,
will be substituted by \code{max(.limit, 100)} if \code{.limit} is set,
otherwise determined by the API (never greater than 100).}

\item{pages}{numeric(1) The number of pages to 'flip' through (default 10)}

\item{org}{character(1L) The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}
}
\value{
A vector of default branches whose names correspond to the
organization's GitHub repositories
}
\description{
This uses the GitHub API, to get the Bioconductor repositories hosted
on GitHub. It returns all packages in https:://github.com/Bioconductor
by default.
}
