% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackagesNEWS.R
\name{getPackagesNEWS}
\alias{getPackagesNEWS}
\alias{printNEWS}
\title{Compare the old and current releases to generate the release announcements
NEWS compilation}
\usage{
getPackagesNEWS(
  prevRepos = "3.15",
  currRepos = "3.16",
  repo = c("bioc", "data/experiment", "workflows"),
  srcdir = NULL
)

printNEWS(
  dbs,
  destfile,
  overwrite = FALSE,
  width = 68,
  output = c("md", "text"),
  relativeLink = FALSE,
  ...
)
}
\arguments{
\item{prevRepos}{character(1) The version string indicating the old release
version of Bioconductor}

\item{currRepos}{character(1) The version string indicating the newest and
current release version of Bioconductor}

\item{repo}{character(1) The repository nickname indicating which repository
to compare news}

\item{srcdir}{(Optional) character(1) The source directory in which all the
Bioconductor packages, whose \code{NEWS} files are to be interrogated, reside}

\item{dbs}{A list of \code{news_db} elements, as returned by \code{getPackageNEWS}.}

\item{destfile}{character(1) file path to the location where NEWS will be
printed.}

\item{overwrite}{logical(1) indicating whether \code{destfile} can be
over-written, if it exists.}

\item{width}{numeric(1) number of characters news items are to be wrapped to,
excluding indent.}

\item{output}{character(1) output to text or markdown format.}

\item{relativeLink}{Should links to packages be relative links on
bioconductor.org website or include full url 'https//bioconductor.org'.
default: FALSE is full url.}

\item{...}{additional arguments, unused.}
}
\value{
A list of NEWS
}
\description{
The function uses previous and current versions of Bioconductor to generate
a single package's \code{NEWS} file.
}
