% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_devel_branch.R
\name{packages_with_default_branch}
\alias{packages_with_default_branch}
\title{Find packages with default branches for an organization}
\usage{
packages_with_default_branch(
  version = BiocManager::version(),
  branches = c(.OLD_DEFAULT_BRANCH, "main"),
  org = "Bioconductor"
)
}
\arguments{
\item{branches}{character() A vector of branches that are sought as default
branches}

\item{org}{character(1L) The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}
}
\value{
A named character vector of default branches whose names correspond
to package repositories on GitHub
}
\description{
This function will search through an organizations repositories and identify
packages whose default branches are in the \code{branches} argument. This allows
the user to identify which repositories will need to have a \code{devel} branch
added.
}
