% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_branch_to_devel.R
\name{rename_branch_repos}
\alias{rename_branch_repos}
\title{Convenience function to create the devel branch for all GitHub repositories}
\usage{
rename_branch_repos(
  repos = character(0L),
  old_branches = c(.OLD_DEFAULT_BRANCH, "main"),
  org = "Bioconductor",
  set_upstream = c("origin/devel", "upstream/devel"),
  clone = TRUE
)
}
\arguments{
\item{repos}{named character() A vector of default branches whose names
correspond to repositories hosted on GitHub. If missing,
\code{repos_with_default_branch} is called and its result is used.}

\item{old_branches}{character() A vector of default branch names to be
replaced, both 'master' and 'main' are included by default. This argument
only works when either \code{packages} or \code{repos} are not specified.}

\item{org}{character(1) The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{set_upstream}{character(1) The remote location that will be tracked by
the local branch, either "origin/devel" (default) or "upstream/devel"}

\item{clone}{logical(1) Whether to clone the GitHub repository into the
current working directory (default: TRUE)}
}
\description{
This function identifies all repositories within an organization that have
\code{old_branches}, i.e., either 'master' or 'main' by default. It then
sets the default branch to \code{devel}.
}
\seealso{
repos_with_default_branch
}
