% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_branch_to_devel.R
\name{repos_with_default_branch}
\alias{repos_with_default_branch}
\title{Identify repositories that have old default branches}
\usage{
repos_with_default_branch(
  branches = c(.OLD_DEFAULT_BRANCH, "main"),
  org = "Bioconductor"
)
}
\arguments{
\item{branches}{character() A vector of branches that are sought as default
branches}

\item{org}{character(1) The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}
}
\value{
A named character vector of default branches whose names correspond
to organization repositories on GitHub
}
\description{
The function obtains all the repositories within the given organization
(Bioconductor) that match the \code{branches} argument.
}
\details{
The output of this function is used to rename branches with
\code{branch_all_repos}.
}
\seealso{
packages_with_default_branch
}
