% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{set_bioc_remote}
\alias{set_bioc_remote}
\title{A convenience function to set the 'upstream' Bioconductor remote}
\usage{
set_bioc_remote(package_name, remote = "upstream")
}
\arguments{
\item{package_name}{named character(1) A scalar string of the default
branch whose name corresponds to a Bioconductor GitHub repository, as
given by \code{packages_list_to_be_updated()}.}

\item{remote}{character(1L) The name of the remote to be created. This is
usually named 'upstream' (default)}
}
\value{
Called for the side effect of creating an 'upstream' remote with the
Bioconductor git address for a given package
}
\description{
The function will create an 'upstream' remote using
\code{git@git.bioconductor.org} as the primary address. If an \code{upstream} remote
already exists, it will be validated. The remote name can be changed to the
desired name via the \code{remote} argument but it is customarily called the
'upstream' remote.
}
