% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_repos.R
\name{sync_repos}
\alias{sync_repos}
\title{Synchronize git repositories from Bioconductor}
\usage{
sync_repos(
  packages,
  dest_dir,
  username,
  org = username,
  set_upstream = "upstream/devel",
  restart_from = character(0L),
  BPPARAM = NULL
)
}
\description{
Either clone or update a list of git repositories (\code{packages})
by name at the desired destination folder (\code{dest_dir}).
}
\examples{
library(BiocPkgTools)
missing_bins <- repositoryStats()[["missing_binaries"]]
sync_repos(missing_bins, "~/bioc/", org = "Bioconductor")

library(BiocParallel)
bpparam <- MulticoreParam(workers = 22)
register(bpparam)

sync_repos(missing_bins, "~/bioc/", org = "Bioconductor", BPPARAM = bpparam)
}
