% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{update_all_packages}
\alias{update_all_packages}
\title{Function to update all the packages.}
\usage{
update_all_packages(release)
}
\arguments{
\item{release}{character(1) The Bioconductor version tag, e.g.,
"RELEASE_3_16"}
}
\description{
Updates all the packages in the GitHub organization maintained by
the core team.
}
