% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_branch_to_devel.R
\name{add_bioc_remote}
\alias{add_bioc_remote}
\title{A convenience function to set the 'upstream' Bioconductor remote}
\usage{
add_bioc_remote(package_name, remote = "upstream")
}
\arguments{
\item{package_name}{character(1) The name of the organization R package that
is also available on GitHub.}

\item{remote}{character(1L) The name of the remote to be created. This is
usually named 'upstream' (default)}
}
\value{
Called for the side effect of creating an 'upstream' remote with the
Bioconductor git address for a given package
}
\description{
The function will create an 'upstream' remote using
\code{git@git.bioconductor.org} as the primary address. If an \code{upstream} remote
already exists, it will be validated. The remote name can be changed to the
desired name via the \code{remote} argument but it is customarily called the
'upstream' remote.
}
