% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_release_gh.R
\name{branch-release-gh}
\alias{branch-release-gh}
\alias{add_gh_release_branch}
\alias{add_gh_release_branches}
\alias{get_bioc_release_yaml}
\title{Add the release branch to GitHub package repositories}
\usage{
add_gh_release_branch(
  package_name,
  release = get_bioc_release_yaml(),
  gh_branch = .BIOC_DEFAULT_BRANCH,
  bioc_branch = .BIOC_DEFAULT_BRANCH,
  org = "Bioconductor"
)

add_gh_release_branches(
  packages = character(0L),
  release = get_bioc_release_yaml(),
  bioc_branch = .BIOC_DEFAULT_BRANCH,
  org = "Bioconductor"
)

get_bioc_release_yaml(config = .BIOC_CONFIG_FILE)
}
\arguments{
\item{package_name}{\code{character(1)} The name of the organization R package
that is also available on GitHub.}

\item{release}{\code{character(1)} The Bioconductor version branch tag, e.g.,
"RELEASE_3_17"}

\item{gh_branch}{\code{character(1)} The name of the default branch on GitHub. It
may be 'devel' or 'main' depending on the repository}

\item{bioc_branch}{\code{character(1)} The name of the default branch on the
Bioconductor git server (default 'devel')}

\item{org}{\code{character(1)} The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{packages}{\verb{named character()} A character vector of default branches
whose names correspond to Bioconductor package names. See
\code{packages_without_release_branch}.}

\item{config}{\code{character(1)} The path to the Bioconductor configuration file
that contains the release version (defaults to website URL from
\code{.BIOC_CONFIG_FILE})}
}
\description{
This function assumes that you have admin push access to the
GitHub organization indicated by \code{org}.
}
\examples{
get_bioc_release_yaml()
if (interactive()) {
    add_gh_release_branch(
      package_name = "BiocParallel",
      release = "RELEASE_3_19"
    )

    add_gh_release_branches(
        release = get_bioc_release_yaml(),
        org = "Bioconductor"
    )
}
}
\seealso{
packages_without_release_branch
}
