% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{branch-release-gh}
\alias{branch-release-gh}
\alias{add_gh_release_branch}
\alias{add_gh_release_branches}
\title{Add the release branch to GitHub package repositories}
\usage{
add_gh_release_branch(
  package_name,
  release = "RELEASE_3_16",
  gh_branch = .BIOC_DEFAULT_BRANCH,
  bioc_branch = .BIOC_DEFAULT_BRANCH,
  org = "Bioconductor"
)

add_gh_release_branches(
  packages = character(0L),
  release = "RELEASE_3_16",
  bioc_branch = .BIOC_DEFAULT_BRANCH,
  org = "Bioconductor"
)
}
\arguments{
\item{package_name}{character(1) The name of the organization R package that
is also available on GitHub.}

\item{release}{character(1) The Bioconductor version branch tag, e.g.,
"RELEASE_3_16"}

\item{gh_branch}{character(1) The name of the default branch on GitHub. It
may be 'main', 'master', or 'devel' depending on the repository}

\item{bioc_branch}{character(1) The name of the default branch on the
Bioconductor git server (default 'master')}

\item{org}{character(1) The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{packages}{named character() A character vector of default branches
whose names correspond to Bioconductor package names. See
\code{packages_without_release_branch}.}
}
\description{
This function assumes that you have admin push access to the GitHub
organization indicated by \code{org}.
}
\examples{
if (interactive()) {
  add_gh_release_branch (
    package_name = "updateObject", gh_branch = "master"
  )
}

}
\seealso{
packages_without_release_branch
}
