% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_release_gh.R
\name{branch-release-gh}
\alias{branch-release-gh}
\alias{add_gh_release_branch}
\alias{add_gh_release_branches}
\title{Add the release branch to GitHub package repositories}
\usage{
add_gh_release_branch(
  package_name,
  release = "RELEASE_3_17",
  gh_branch = .BIOC_DEFAULT_BRANCH,
  bioc_branch = .BIOC_DEFAULT_BRANCH,
  org = "Bioconductor"
)

add_gh_release_branches(
  packages = character(0L),
  release = "RELEASE_3_17",
  bioc_branch = .BIOC_DEFAULT_BRANCH,
  org = "Bioconductor"
)
}
\arguments{
\item{package_name}{\code{character(1)} The name of the organization R package that
is also available on GitHub.}

\item{release}{\code{character(1)} The Bioconductor version branch tag, e.g.,
"RELEASE_3_17"}

\item{gh_branch}{\code{character(1)} The name of the default branch on GitHub. It
may be 'devel' or 'main' depending on the repository}

\item{bioc_branch}{\code{character(1)} The name of the default branch on the
Bioconductor git server (default 'devel')}

\item{org}{\code{character(1)} The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{packages}{\verb{named character()} A character vector of default branches
whose names correspond to Bioconductor package names. See
\code{packages_without_release_branch}.}
}
\description{
This function assumes that you have admin push access to the GitHub
organization indicated by \code{org}.
}
\examples{
if (interactive()) {
  add_gh_release_branch (
    package_name = "updateObject", gh_branch = "devel"
  )
}

}
\seealso{
packages_without_release_branch
}
