% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{clone_and_push_git_repo}
\alias{clone_and_push_git_repo}
\title{Clone and update a GitHub repository.}
\usage{
clone_and_push_git_repo(
  package_name,
  release = "RELEASE_3_16",
  gh_branch = "master",
  bioc_branch = "master",
  org = "Bioconductor"
)
}
\arguments{
\item{package_name}{named character(1) A scalar string of the default
branch whose name corresponds to a Bioconductor GitHub repository, as
given by \code{packages_list_to_be_updated()}.}

\item{release}{character(1) The Bioconductor version branch tag, e.g.,
"RELEASE_3_16"}

\item{gh_branch}{character(1) The name of the default branch on GitHub. It
may be 'main', 'master', or 'devel' depending on the repository}

\item{bioc_branch}{character(1) The name of the default branch on the
Bioconductor git server (default 'master')}
}
\description{
This function assumes that you have admin push access to the
bioconductor github organization.
}
\examples{
if (interactive()) {
  clone_and_push_git_repo(
    package_name = "updateObject", gh_branch = "master"
  )
}

}
