% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_release_gh.R
\name{get-github-repos}
\alias{get-github-repos}
\alias{get_org_github_repos}
\alias{get_user_github_repos}
\title{Get all repositories for a given GitHub organization or username}
\usage{
get_org_github_repos(
  per_page = 100,
  pages = 10,
  org = "Bioconductor",
  archived = FALSE
)

get_user_github_repos(per_page = 100, pages = 10, username, archived = FALSE)
}
\arguments{
\item{per_page}{Number of items to return per page. If omitted,
will be substituted by \code{max(.limit, 100)} if \code{.limit} is set,
otherwise determined by the API (never greater than 100).}

\item{pages}{numeric(1) The number of pages to 'flip' through (default 10)}

\item{org}{character(1) The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{archived}{logical(1) Whether to include archived repositories in the
query results (default FALSE)}

\item{username}{character(1) The GitHub username used to query repositories}
}
\value{
A vector of default branches whose names correspond to the
organization or user GitHub repositories
}
\description{
\code{get_user_github_repos} and \code{get_org_github_repos} use the
GitHub API, to obtain the repositories hosted on GitHub via the 'username'
or the organization ('org') names, respectively.
}
\examples{

if (interactive()) {
  my_repos <- get_user_github_repos(username = "github-username")
  change <- my_repos[my_repos == "master"]
  if (length(change))
    rename_branch_repos(
      repos = change,
      org = "github-username",
      clone = TRUE
    )
}

}
