% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{get_bioc_software_manifest}
\alias{get_bioc_software_manifest}
\title{Get bioconductor software manifest}
\usage{
get_bioc_software_manifest()
}
\description{
The software packages are the only packages hosted on github, by
the organization.
}
