% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackagesNEWS.R
\name{getPackagesNEWS}
\alias{getPackagesNEWS}
\alias{printNEWS}
\title{Compare the old and current releases to generate the release announcements
NEWS compilation}
\usage{
getPackagesNEWS(
  prevRepos = "3.15",
  currRepos = "3.16",
  repo = c("bioc", "data/experiment", "workflows"),
  srcdir = NULL
)

printNEWS(
  dbs,
  destfile,
  overwrite = FALSE,
  width = 68,
  output = c("md", "text"),
  relativeLink = FALSE,
  ...
)
}
\arguments{
\item{prevRepos}{\code{character(1)} The version string indicating the old release
version of Bioconductor}

\item{currRepos}{\code{character(1)} The version string indicating the newest and
current release version of Bioconductor}

\item{repo}{\code{character(1)} The repository nickname indicating which repository
to compare news}

\item{srcdir}{(Optional) \code{character(1)} The source directory in which all the
Bioconductor packages, whose \code{NEWS} files are to be interrogated, reside}

\item{dbs}{\code{list()} The \code{news_db} list as returned by \code{getPackageNEWS}.}

\item{destfile}{\code{character(1)} The file path location to where \code{NEWS} will be
saved.}

\item{overwrite}{\code{logical(1)} Indicate whether \code{destfile} should be
over-written, if it exists.}

\item{width}{\code{numeric(1)} The maximum number of characters \code{NEWS} items can
span before wrapping, excludes indent.}

\item{output}{\code{character(1)} Either 'md' or 'text' indicating the output in
markdown or text format, respectively.}

\item{relativeLink}{\code{logical(1)} Whether to include links relative to the
Bioconductor website or provide the full URL, i.e.,
\url{https://bioconductor.org}  (default: \code{FALSE}).}

\item{...}{additional arguments, unused.}
}
\value{
A list of NEWS
}
\description{
The function uses previous and current versions of Bioconductor to generate
a single package's \code{NEWS} file.
}
