% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackagesNEWS.R
\name{getPackagesNEWS}
\alias{getPackagesNEWS}
\title{Compare the old and current releases to generate the release announcements
NEWS compilation}
\usage{
getPackagesNEWS(
  prevRepos = "3.15",
  currRepos = "3.16",
  repo = c("bioc", "data/experiment", "workflows"),
  srcdir = NULL
)
}
\arguments{
\item{prevRepos}{character(1) The version string indicating the old release
version of Bioconductor}

\item{currRepos}{character(1) The version string indicating the newest and
current release version of Bioconductor}

\item{repo}{character(1) The repository nickname indicating which repository
to compare news}

\item{srcdir}{(Optional) character(1) The source directory in which all the
Bioconductor packages, whose \code{NEWS} files are to be interrogated, reside}
}
\value{
A list of NEWS
}
\description{
The function uses previous and current versions of Bioconductor to generate
a single package's \code{NEWS} file.
}
