% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{packages_list_to_be_updated}
\alias{packages_list_to_be_updated}
\title{Generate list of packages to be updated}
\usage{
packages_list_to_be_updated(version = "3.16")
}
\description{
Generate list of packages to be updated
}
