% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_packages.R
\name{packages_list_to_be_updated}
\alias{packages_list_to_be_updated}
\title{Generate the list of packages to be updated}
\usage{
packages_list_to_be_updated(
  version = "3.16",
  org = "Bioconductor",
  type = c("BioCsoft", "BioCexp")
)
}
\arguments{
\item{version}{character(1L) The numeric version of the Bioconductor release,
e.g., "3.16"}

\item{org}{character(1) The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{type}{character(1L) The official repository name as given by
\code{BiocManager::repositories()}. Currently, only software and experiment data
('BioCsoft' and 'BioCexp', respectively) are supported.}
}
\value{
A named scalar string of the default branch whose name corresponds to
a Bioconductor GitHub repository
}
\description{
This function obtains all the repositories from the Bioconductor organization
and filters them to only valid R packages and repositories that do not have a
\code{RELEASE_XX_YY} branch.
}
