% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch_release_gh.R
\name{packages_without_release_branch}
\alias{packages_without_release_branch}
\title{Generate the list of packages to be updated}
\usage{
packages_without_release_branch(
  version = "3.16",
  org = "Bioconductor",
  type = c("BioCsoft", "BioCexp")
)
}
\arguments{
\item{version}{\code{character(1)} The numeric version of the Bioconductor release,
e.g., "3.16"}

\item{org}{\code{character(1)} The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{type}{\code{character(1)} The official repository name as given by
\code{BiocManager::repositories()}. Currently, only software and experiment data
('BioCsoft' and 'BioCexp', respectively) are supported.}
}
\value{
A named scalar string of the default branch whose name corresponds to
a Bioconductor GitHub repository
}
\description{
This function obtains all the repositories from the Bioconductor organization
and filters them to only valid R packages and repositories that do not have a
\code{RELEASE_XX_YY} branch.
}
