% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_local_repos.R
\name{update_local_repos}
\alias{update_local_repos}
\alias{update_local_repo}
\title{Update local repositories after Bioconductor Release}
\usage{
update_local_repos(
  repos_dir,
  release = bioc_release_yaml(),
  username,
  org = username,
  set_upstream = "origin/devel",
  restart_from = character(0L),
  BPPARAM = NULL
)

update_local_repo(
  repo_dir,
  release,
  set_upstream = "origin/devel",
  org = "Bioconductor"
)
}
\arguments{
\item{repos_dir}{\code{character(1)} The base directory where all packages /
repositories exist for the user}

\item{release}{\code{character(1)} The Bioconductor release version, as
"RELASE_X_Y", to use for updating the local repositories. By default, the
value of \code{bioc_release_yaml()}.}

\item{username}{\code{character(1)} (optional) The GitHub username used in the
query to check default packages}

\item{org}{\code{character(1)} The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{set_upstream}{\code{character(1)} The remote location that will be tracked
by the local branch, either "origin/devel" (default) or "upstream/devel"}

\item{restart_from}{\code{character(1)} The name of a package found in \code{repos_dir}
that should be used as the starting point for updating the local repos in
case that the process was interrupted due to merge conflicts etc. Default
\code{character(0L)}.}

\item{BPPARAM}{\code{BiocParallelParam} An optional \code{BiocParallelParam} instance
defining the parallel back-end to be used during evaluation. Default
\code{NULL}.}

\item{repo_dir}{\code{character(1)} The full path to a single package / repository
whose default branch should be updated}

\item{packages}{\code{character()} An optional vector of package names that
correspond to folders available locally.}
}
\description{
These functions should be used after a Bioconductor Release. They ensure that
the local repositories are in sync with Bioconductor. For convenience, the
singular \code{update_local_repo} function will update a single local repository
on the user's system. The function checks repositories against the
\code{available.packages()} in the Bioconductor owned
\code{BiocManager::repositories()}.
}
\details{
The \code{update_local_repos} function will update all local repositories
in the \code{repos_dir} directory that are found in either the users or
organization's GitHub repository. The function syncs the \code{devel} and
current release branches from both the \code{origin} (GH) and \code{upstream}
(Bioconductor) remotes. If a \code{upstream} remote is not set, the function
will set it to the Bioconductor git server. Finally, the function sets
\code{origin/devel} as the remote branch to track with the \code{set_upstream}
argument.
}
\examples{
if (interactive()) {
    ## add PAT credentials
    gitcreds::gitcreds_set()
    ## update multiple packages at a time
    update_local_repos(
        repos_dir = "~/bioc/",
        org = "Bioconductor",
        release = "RELEASE_3_20"
    )

    ## update a single package
    update_local_repo(
        "~/bioc/AnnotationHub",
        release = bioc_release_yaml(),
        org = "Bioconductor"
    )
    setwd("~/bioc/AnnotationHub")
    update_local_repo(
        ".",
        release = bioc_release_yaml(),
        org = "Bioconductor"
    )
}
}
