% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_local_repos.R
\name{update_local_repos}
\alias{update_local_repos}
\alias{update_local_repo}
\title{Update local repositories after Bioconductor Release}
\usage{
update_local_repos(
  repos_dir,
  release = bioc_release_yaml(),
  username,
  org = username,
  set_upstream = "origin/devel"
)

update_local_repo(
  repo_dir,
  release,
  set_upstream = "origin/devel",
  org = "Bioconductor"
)
}
\arguments{
\item{repos_dir}{\code{character(1)} The base directory where all packages /
repositories exist for the user}

\item{release}{\code{character(1)} The Bioconductor release version, as
"RELASE_X_Y", to use for updating the local repositories. By default, the
value of \code{bioc_release_yaml()}.}

\item{username}{\code{character(1)} (optional) The GitHub username used in the
query to check default packages}

\item{org}{\code{character(1)} The organization for which to extract the names of
the repositories on GitHub (default "Bioconductor").}

\item{set_upstream}{\code{character(1)} The remote location that will be tracked
by the local branch, either "origin/devel" (default) or "upstream/devel"}

\item{repo_dir}{\code{character(1)} The full path to a single package / repository
whose default branch should be updated}

\item{packages}{\code{character()} An optional vector of package names that
correspond to folders available locally.}
}
\description{
These functions should be used after a Bioconductor Release. They ensure that
the local repositories are in sync with Bioconductor. For convenience, the
singular \code{update_local_repo} function will update a single local repository
on the user's system.
}
\examples{
if (interactive()) {
    ## update multiple packages at a time
    update_local_repos(repos_dir = "~/bioc/", org = "Bioconductor")
    update_local_repos(
        packages = c("~/bioc/AnnotationHub", "~/bioc/BiocGenerics"),
        org = "Bioconductor"
    )
    update_local_repos(
        repos_dir = "~/bioc/", packages = c("AnnotationHub", "BiocGenerics"),
        org = "Bioconductor"
    )

    ## update a single package
    update_local_repo(
        "~/bioc/AnnotationHub",
        release = get_bioc_release_yaml(),
        org = "Bioconductor"
    )
    setwd("~/bioc/AnnotationHub")
    update_local_repo(
        ".",
        release = get_bioc_release_yaml(),
        org = "Bioconductor"
    )
}
}
